More refactoring and improvements: ConnectAuth is the main class and methods are static, common types were moved to types.ts, update readme
