+++ src/index.ts
@@ -1,5 +1,5 @@
 export * from './AuthError';
-export * from './ConnectApp';
+export * from './ConnectAuth';
 export * from './Jwt';
 export * from './requestReader';
 export * from './types';
