+++ src/types.ts
@@ -2,10 +2,9 @@ type Timestamp = number;
 
 export const ContextQsh = 'context-qsh';
 
-export interface ConnectCredentials {
-  sharedSecret: string;
-}
-
+/**
+ * Represents an extensible payload of a Connect JWT.
+ */
 export interface ConnectJwt {
   sub: string;
   qsh?: string | typeof ContextQsh;
@@ -15,4 +14,24 @@ export interface ConnectJwt {
   iat: Timestamp;
   context: unknown;
   [key: string]: unknown;
+}
+
+export enum InstallType {
+  newInstall = 'newInstall',
+  update = 'update',
+}
+
+/**
+ * Holds current credentials and the entity representing an existing installation.
+ */
+export interface CredentialsWithEntity<E> {
+  sharedSecret: string;
+  storedEntity: E;
+}
+
+/**
+ * Used to load data from an existing installation.
+ */
+export interface CredentialsLoader<E> {
+  (clientKey: string): Promise<CredentialsWithEntity<E> | undefined>;
 }
