+++ test/verifyInstall.test.ts
@@ -1,67 +1,73 @@
 import * as atlassianJwt from 'atlassian-jwt';
 
-import { AuthError, AuthErrorCode, ConnectApp, ConnectJwt, InstallType } from '../src';
+import { AuthError, AuthErrorCode, ConnectAuth, ConnectJwt, InstallType } from '../src';
 import { TestRequestReader } from './helpers/util';
 
 const baseUrl = 'https://test.example.com';
 const jiraClientKey = 'jira-client-key';
 const bitbucketClientKey = 'bitbucket-client-key';
-const entity = { appSpecificId: 1 };
+const storedEntity = { appSpecificId: 1 };
 
 const jiraPayload = {
   baseUrl: 'https://test.atlassian.net',
   clientKey: jiraClientKey,
   sharedSecret: 'shh-secret-cat',
-  entity,
+  storedEntity,
 };
+
 const bitbucketPayload = {
   principal: { uuid: 'bitbucket-workspace-id' },
   clientKey: bitbucketClientKey,
   sharedSecret: 'shh-secret-cat',
-  entity,
+  storedEntity,
 };
 
-const jiraApp = new ConnectApp({
-  baseUrl,
-  checkQueryStringHashOnInstall: true,
-  checkQueryStringHashOnRequest: true,
-});
-const bitbucketApp = new ConnectApp({
-  baseUrl,
-  checkQueryStringHashOnInstall: false,
-  checkQueryStringHashOnRequest: false,
-});
+describe('ConnectAuth.verifyInstall', () => {
+  const loadCredentials = jest.fn();
 
-describe('ConnectApp.verifyInstall', () => {
-  const table = [
-    { name: 'Jira', app: jiraApp, clientKey: jiraClientKey },
-    { name: 'Bitbucket', app: bitbucketApp, clientKey: bitbucketClientKey },
-  ];
+  beforeEach(() => jest.clearAllMocks());
 
-  describe.each(table)('$name app', ({ app, clientKey }) => {
-    test('First install', async () => {
-      const requestReader = new TestRequestReader({ qsh: '', clientKey, jwt: '' });
-      const loadCredentials = jest.fn();
+  test('First install', async () => {
+    const clientKey = jiraClientKey;
+    const requestReader = new TestRequestReader({ qsh: '', clientKey, jwt: '' });
 
-      const result = await app.verifyInstall({ requestReader, loadCredentials });
+    const result = await ConnectAuth.verifyInstall({ baseUrl, requestReader, loadCredentials });
 
-      expect(result).toStrictEqual({
-        type: InstallType.newInstall,
-        clientKey,
-      });
-      expect(loadCredentials).toHaveBeenCalledWith(clientKey);
+    expect(result).toStrictEqual({
+      type: InstallType.newInstall,
+      clientKey,
     });
+    expect(loadCredentials).toHaveBeenCalledWith(clientKey);
+  });
+
+  test('First install without QSH checking', async () => {
+    const clientKey = bitbucketClientKey;
+    const requestReader = new TestRequestReader({ qsh: '', clientKey, jwt: '' });
+
+    const result = await ConnectAuth.verifyInstall({
+      baseUrl,
+      requestReader,
+      loadCredentials,
+      skipQueryStringHashCheck: true,
+    });
+
+    expect(result).toStrictEqual({
+      type: InstallType.newInstall,
+      clientKey,
+    });
+    expect(loadCredentials).toHaveBeenCalledWith(clientKey);
   });
 
   test('Failed to decode token', async () => {
-    const loadCredentials = jest.fn();
     const requestReader = new TestRequestReader({
       qsh: '',
       clientKey: jiraClientKey,
       jwt: 'abc.def.ghi',
     });
 
-    await expect(jiraApp.verifyInstall({ requestReader, loadCredentials })).rejects.toMatchError(
+    await expect(
+      ConnectAuth.verifyInstall({ baseUrl, requestReader, loadCredentials })
+    ).rejects.toMatchError(
       new AuthError('Failed to decode token', {
         code: AuthErrorCode.FAILED_TO_DECODE,
         originError: new SyntaxError('Unexpected token i in JSON at position 0'),
@@ -70,7 +76,6 @@ describe('ConnectApp.verifyInstall', () => {
   });
 
   test('Passed different id in body and authorization header', async () => {
-    const loadCredentials = jest.fn();
     const clientKey = jiraClientKey;
     const jwtContent = { iss: 'different-id' } as unknown as ConnectJwt;
     const jwt = atlassianJwt.encodeSymmetric(jwtContent, jiraPayload.sharedSecret);
@@ -80,7 +85,9 @@ describe('ConnectApp.verifyInstall', () => {
       jwt,
     });
 
-    await expect(jiraApp.verifyInstall({ requestReader, loadCredentials })).rejects.toMatchError(
+    await expect(
+      ConnectAuth.verifyInstall({ baseUrl, requestReader, loadCredentials })
+    ).rejects.toMatchError(
       new AuthError('Wrong issuer', {
         code: AuthErrorCode.WRONG_ISSUER,
         unverifiedConnectJwt: jwtContent,
@@ -91,7 +98,7 @@ describe('ConnectApp.verifyInstall', () => {
   });
 
   test('Second and subsequent installation of Jira add-on with no qsh', async () => {
-    const loadCredentials = jest.fn().mockReturnValue(jiraPayload);
+    loadCredentials.mockReturnValue(jiraPayload);
     const clientKey = jiraClientKey;
     const jwtContent = { iss: jiraPayload.clientKey } as unknown as ConnectJwt;
     const jwt = atlassianJwt.encodeSymmetric(jwtContent, jiraPayload.sharedSecret);
@@ -101,7 +108,9 @@ describe('ConnectApp.verifyInstall', () => {
       jwt,
     });
 
-    await expect(jiraApp.verifyInstall({ requestReader, loadCredentials })).rejects.toMatchError(
+    await expect(
+      ConnectAuth.verifyInstall({ baseUrl, requestReader, loadCredentials })
+    ).rejects.toMatchError(
       new AuthError('JWT did not contain the query string hash (qsh) claim', {
         code: AuthErrorCode.MISSING_QSH,
         connectJwt: jwtContent,
@@ -110,7 +119,7 @@ describe('ConnectApp.verifyInstall', () => {
   });
 
   test('Second and subsequent installation of Bitbucket add-on with no qsh', async () => {
-    const loadCredentials = jest.fn().mockReturnValue(bitbucketPayload);
+    loadCredentials.mockReturnValue(bitbucketPayload);
     const clientKey = bitbucketClientKey;
     const jwtContent = {
       iss: bitbucketPayload.clientKey,
@@ -122,19 +131,24 @@ describe('ConnectApp.verifyInstall', () => {
       jwt,
     });
 
-    const result = await bitbucketApp.verifyInstall({ requestReader, loadCredentials });
+    const result = await ConnectAuth.verifyInstall({
+      baseUrl,
+      requestReader,
+      loadCredentials,
+      skipQueryStringHashCheck: true,
+    });
 
     expect(result).toStrictEqual({
       type: InstallType.update,
       clientKey,
       connectJwt: jwtContent,
-      entity,
+      storedEntity,
     });
     expect(loadCredentials).toHaveBeenCalledWith(clientKey);
   });
 
   test('Second and subsequent Jira add-on install', async () => {
-    const loadCredentials = jest.fn().mockReturnValue(jiraPayload);
+    loadCredentials.mockReturnValue(jiraPayload);
     const clientKey = jiraClientKey;
     const qsh = 'valid';
     const jwtContent = {
@@ -148,19 +162,19 @@ describe('ConnectApp.verifyInstall', () => {
       jwt,
     });
 
-    const result = await jiraApp.verifyInstall({ requestReader, loadCredentials });
+    const result = await ConnectAuth.verifyInstall({ baseUrl, requestReader, loadCredentials });
 
     expect(result).toStrictEqual({
       type: InstallType.update,
       clientKey,
       connectJwt: jwtContent,
-      entity,
+      storedEntity,
     });
     expect(loadCredentials).toHaveBeenCalledWith(clientKey);
   });
 
   test('Unauthorized request to updated existing instance', async () => {
-    const loadCredentials = jest.fn().mockReturnValue(jiraPayload);
+    loadCredentials.mockReturnValue(jiraPayload);
     const clientKey = jiraClientKey;
     const requestReader = new TestRequestReader({
       qsh: '',
@@ -168,7 +182,9 @@ describe('ConnectApp.verifyInstall', () => {
       jwt: '',
     });
 
-    await expect(jiraApp.verifyInstall({ requestReader, loadCredentials })).rejects.toMatchError(
+    await expect(
+      ConnectAuth.verifyInstall({ baseUrl, requestReader, loadCredentials })
+    ).rejects.toMatchError(
       new AuthError('Unauthorized update request', { code: AuthErrorCode.UNAUTHORIZED_REQUEST })
     );
     expect(loadCredentials).toHaveBeenCalledWith(clientKey);
